vrp_validation/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
folium>=0.12.0
pulp>=2.6.0
ortools>=9.4.0
pytest>=7.0.0
jupyter>=1.0.0
pyyaml>=6.0.0